requirements.txt
Pillow
pyinstaller
matplotlib
